refactor: rename handlers to view
