+++ src/main/kotlin/lotto/Application.kt
@@ -1,17 +1,17 @@
 package lotto
 
-import lotto.handler.InputHandler
-import lotto.handler.OutputHandler
+import lotto.view.InputView
+import lotto.view.OutputView
 import lotto.service.LottoManager
 
 fun main() {
     // TODO: Implement the program
-    val purchasedAmount = InputHandler.getPurchaseAmount();
+    val purchasedAmount = InputView.getPurchaseAmount();
 
     val ticketAmount = purchasedAmount / 1000
 
     val tickets = LottoManager.generateMultipleTickets(ticketAmount)
 
-    val outputHandler = lotto.handler.OutputHandler()
-    outputHandler.displayPurchasedTicketsInfo(tickets)
+    val outputView = lotto.view.OutputView()
+    outputView.displayPurchasedTicketsInfo(tickets)
 }
